Refactor SeatTypeViewer into TypeScript with modular components and hooks
